DSP: Do not generate warnings when scalar functions are used with MVE
This commit updates the DSP functions to print out a compiler message
(non-warning) instead of a warning when the scalar version of a
function implementation is used in spite of the M-Profile Vector
Extension being enabled.

The rationale is that warnings may be treated as errors under certain
circumstances (e.g. when running tests), and the scalar function
implementations being used should not cause a buiuld failure.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
GitOrigin-RevId: 100147e8ba9b42e1da1f5732078c32d08182441c
Change-Id: I9e0e87f04ffc5202e62fb5e5e643411ae25fe330
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/cmsis/+/3627932
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>